# product/src/main/java/com/ecom/product/dto/CategoryRequest.java
package com.ecom.product.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class CategoryRequest {
    @NotBlank
    private String name;
}