# user/src/main/java/com/ecom/user/dto/UserResponse.java
package com.ecom.user.dto;

import com.ecom.user.model.UserRole;
import lombok.Data;

@Data
public class UserResponse {
    private String id;
    private String firstName;
    private String lastName;
    private String email;
    private String phone;
    private UserRole role;
    private AddressDTO address;
}